refactor: new io module adopting bioio
